+++ src/panel/ExternalCommandData.hh
@@ -1,6 +1,6 @@
 //
 // ExternalCommandData.hh for pekwm
-// Copyright (C) 2022 Claes Nästén <pekdon@gmail.com>
+// Copyright (C) 2022-2025 Claes Nästén <pekdon@gmail.com>
 //
 // This program is licensed under the GNU GPL.
 // See the LICENSE file for more information.
@@ -31,25 +31,28 @@ public:
 	class CommandProcess
 	{
 	public:
-		CommandProcess(const std::string& command, uint interval_s);
-		~CommandProcess(void);
+		CommandProcess(const std::string& command, uint interval_s,
+			       const std::string& assign);
+		~CommandProcess();
 
-		int getFd(void) const { return _fd; }
-		pid_t getPid(void) const { return _pid; }
-		std::string& getBuf(void) { return _buf; }
+		int getFd() const { return _fd; }
+		pid_t getPid() const { return _pid; }
+		std::string& getBuf() { return _buf; }
+		const std::string& getAssign() const { return _assign; }
 
-		bool start(void);
+		bool start();
 
 		bool checkInterval(struct timespec *now)
 		{
 			return now->tv_sec >= _next_interval.tv_sec;
 		}
 
-		void reset(void);
+		void reset();
 
 	private:
 		std::string _command;
 		uint _interval_s;
+		std::string _assign;
 		struct timespec _next_interval;
 
 		pid_t _pid;
@@ -58,14 +61,15 @@ public:
 	};
 
 	ExternalCommandData(const PanelConfig& cfg, VarData& var_data);
-	~ExternalCommandData(void);
+	virtual ~ExternalCommandData();
 
 	void refresh(fdFun addFd, void *opaque);
 	bool input(int fd);
 	void done(pid_t pid, fdFun removeFd, void *opaque);
 
-private:
-	void append(std::string &buf, char *data, size_t size);
+protected:
+	void append(std::string &buf, const char *data, size_t size,
+		    const std::string &assign);
 	void parseOutput(const std::string& buf);
 	void parseLine(const std::string& line);