+++ test/test_pekwm_panel.cc
@@ -1,12 +1,14 @@
 //
 // test_pekwm_panel.cc for pekwm
-// Copyright (C) 2021-2023 Claes Nästén <pekdon@gmail.com>
+// Copyright (C) 2021-2025 Claes Nästén <pekdon@gmail.com>
 //
 // This program is licensed under the GNU GPL.
 // See the LICENSE file for more information.
 //
 
 #include "test.hh"
+
+#include "test_ExternalCommandData.hh"
 
 #include "pekwm.hh"
 #include "Debug.hh"
@@ -18,6 +20,8 @@ main_tests(int argc, char *argv[])
 	// Setup environment required for the tests
 	Debug::setLogFile("/dev/null");
 	X11::addHead(Head(0, 0, 800, 600));
+
+	TestExternalCommandData externalCommandData;
 
 	return TestSuite::main(argc, argv);
 }